Move Runner to Dispatcher.
Since calls will be happen in the caller process, we don't actually need a supervisor set up to run things in separate processes.  We just need our dispatcher to be synchronized in doling out the function pointer, etc, so that we can return when things are tripped, etc.  That should be fast enough, since it's really just map manipulation, and lets us avoid the burden of running processes ourselves.

The bummer is that I can't think of a good way to return back to a caller when something hits a timeout, since right now, we'll track latency thresholds being crossed but the calls will still take forever when the caller runs them.  Gotta think on that one...