Start to flesh out the logic of the dispatcher.
I'm taking a linear scan approach for the first cut, where we'll keep a list of good/bad requests and when they happened.  When we add a new request -- good or bad-- to the list, we'll prune the entire list to make sure we only have requests from the last N seconds.  When I flesh it out, the request to run a function will again prune the list to make sure it's current, and then compute the error percentage / average latency of the window right then and there.  For moderate request rates, this should be OK but it's definitely not optimal.

I'm considering some sort of background process that culls the list every second or so and keeps track of whether the circuit should be tripped or not, which would make it O(1) for callers.  Right now, though, I just want to make sure everything is sound.