Rename 'functions' to 'commands.
This matches the Hystrix parlance, but it also is just less confusing in general.  A command constitutes a named object where there's a function to run and stats about its historical invocations.  We have the boilerplate to start figuring out if a command is tripped or not, but right now circuits are hard 'closed'.