First cut of an opportunistic command update mechanism.
Instead of the update-every-request model, or the update-every-second model, I'm trying out a threshold-based update mechanism.  Every request, we check to see if the command's status needs to be updated.  This happens if we haven't updated within 500ms, but might eventually be tweaked to be a dynamic time period or every N requests.  The tests are a little slower because we have to sleep, but still pretty fast.